Support my imports log scroll to bottom behavior
